Extract full navigation width header style
